MetadataServicesController (note: tests changed to reflect the fact that 'name' should be required)
