+++ api/admin/routes.py
@@ -438,14 +438,14 @@ def reset_adobe_id():
 @requires_admin
 @requires_csrf_token
 def metadata_services():
-    return app.manager.admin_settings_controller.metadata_services()
+    return app.manager.admin_metadata_services_controller.process_metadata_services()
 
 @app.route("/admin/metadata_service/<service_id>", methods=["DELETE"])
 @returns_json_or_response_or_problem_detail
 @requires_admin
 @requires_csrf_token
 def metadata_service(service_id):
-    return app.manager.admin_settings_controller.metadata_service(service_id)
+    return app.manager.admin_metadata_services_controller.process_delete(service_id)
 
 @app.route("/admin/analytics_services", methods=['GET', 'POST'])
 @returns_json_or_response_or_problem_detail
